Add new module to avoid duplicate code
