+++ Examples/Tests/collision/analysis_collision_3d.py
@@ -28,6 +28,7 @@ import yt
 import math
 import numpy
 from glob import glob
+import post_processing_utils
 sys.path.insert(1, '../../../../warpx/Regression/Checksum/')
 import checksumAPI
 
@@ -80,111 +81,23 @@ assert(error < tolerance)
 ## In the second past of the test, we verify that the diagnostic particle filter function works as
 ## expected. For this, we only use the last simulation timestep.
 
-## This is a generic function to test a particle filter. We reproduce the filter in python and
-## verify that the results are the same as with the filtered diagnostic.
-def check_particle_filter(fn, filtered_fn, filter_expression):
-    ds  = yt.load( fn )
-    ds_filtered  = yt.load( filtered_fn )
-    ad  = ds.all_data()
-    ad_filtered  = ds_filtered.all_data()
-
-    ## Load arrays from the unfiltered diagnostic
-    ids = ad['electron', 'particle_id'].to_ndarray()
-    cpus = ad['electron', 'particle_cpu'].to_ndarray()
-    x = ad['electron', 'particle_position_x'].to_ndarray()
-    px  = ad['electron', 'particle_momentum_x'].to_ndarray()
-    y = ad['electron', 'particle_position_y'].to_ndarray()
-    py  = ad['electron', 'particle_momentum_y'].to_ndarray()
-    z = ad['electron', 'particle_position_z'].to_ndarray()
-    pz  = ad['electron', 'particle_momentum_z'].to_ndarray()
-    w  = ad['electron', 'particle_weight'].to_ndarray()
-
-    ## Load arrays from the filtered diagnostic
-    ids_filtered_warpx = ad_filtered['particle_id'].to_ndarray()
-    cpus_filtered_warpx = ad_filtered['particle_cpu'].to_ndarray()
-    x_filtered_warpx = ad_filtered['particle_position_x'].to_ndarray()
-    px_filtered_warpx  = ad_filtered['particle_momentum_x'].to_ndarray()
-    y_filtered_warpx  = ad_filtered['particle_position_y'].to_ndarray()
-    py_filtered_warpx  = ad_filtered['particle_momentum_y'].to_ndarray()
-    z_filtered_warpx = ad_filtered['particle_position_z'].to_ndarray()
-    pz_filtered_warpx  = ad_filtered['particle_momentum_z'].to_ndarray()
-    w_filtered_warpx  = ad_filtered['particle_weight'].to_ndarray()
-
-    ## Reproduce the filter in python: this returns the indices of the filtered particles in the
-    ## unfiltered arrays.
-    ind_filtered_python, = numpy.where(eval(filter_expression))
-
-    ## Sort the indices of the filtered arrays by particle id.
-    sorted_ind_filtered_python = ind_filtered_python[numpy.argsort(ids[ind_filtered_python])]
-    sorted_ind_filtered_warpx = numpy.argsort(ids_filtered_warpx)
-
-    ## Check that the sorted ids are exactly the same with the warpx filter and the filter
-    ## reproduced in python
-    assert(numpy.array_equal(ids[sorted_ind_filtered_python],
-                             ids_filtered_warpx[sorted_ind_filtered_warpx]))
-    assert(numpy.array_equal(cpus[sorted_ind_filtered_python],
-                             cpus_filtered_warpx[sorted_ind_filtered_warpx]))
-
-    ## Finally, we check that the sum of the particles quantities are the same to machine precision
-    tolerance_checksum = 1.e-12
-    check_array_sum(x[sorted_ind_filtered_python],
-                    x_filtered_warpx[sorted_ind_filtered_warpx], tolerance_checksum)
-    check_array_sum(px[sorted_ind_filtered_python],
-                    px_filtered_warpx[sorted_ind_filtered_warpx], tolerance_checksum)
-    check_array_sum(y[sorted_ind_filtered_python],
-                    y_filtered_warpx[sorted_ind_filtered_warpx], tolerance_checksum)
-    check_array_sum(py[sorted_ind_filtered_python],
-                    py_filtered_warpx[sorted_ind_filtered_warpx], tolerance_checksum)
-    check_array_sum(z[sorted_ind_filtered_python],
-                    z_filtered_warpx[sorted_ind_filtered_warpx], tolerance_checksum)
-    check_array_sum(pz[sorted_ind_filtered_python],
-                    pz_filtered_warpx[sorted_ind_filtered_warpx], tolerance_checksum)
-    check_array_sum(w[sorted_ind_filtered_python],
-                    w_filtered_warpx[sorted_ind_filtered_warpx], tolerance_checksum)
-
-## This function checks that the absolute sums of two arrays are the same to a required precision
-def check_array_sum(array1, array2, tolerance_checksum):
-    sum1 = numpy.sum(numpy.abs(array1))
-    sum2 = numpy.sum(numpy.abs(array2))
-    assert(abs(sum2-sum1)/sum1 < tolerance_checksum)
-
-## This function is specifically used to test the random filter. First, we check that the number of
-## dumped particles is as expected. Next, we call the generic check_particle_filter function.
-def check_random_filter(fn, filtered_fn, random_fraction):
-    ds  = yt.load( fn )
-    ds_filtered  = yt.load( filtered_fn )
-    ad  = ds.all_data()
-    ad_filtered  = ds_filtered.all_data()
-
-    ## Check that the number of particles is as expected
-    numparts = ad['electron', 'particle_id'].to_ndarray().shape[0]
-    numparts_filtered = ad_filtered['particle_id'].to_ndarray().shape[0]
-    expected_numparts_filtered = random_fraction*numparts
-    # 5 sigma test that has an intrinsic probability to fail of 1 over ~2 millions
-    std_numparts_filtered = numpy.sqrt(expected_numparts_filtered)
-    error = abs(numparts_filtered-expected_numparts_filtered)
-    print("Random filter: difference between expected and actual number of dumped particles: " \
-          + str(error))
-    print("tolerance: " + str(5*std_numparts_filtered))
-    assert(error<5*std_numparts_filtered)
-
-    ## Dirty trick to find particles with the same ID + same CPU (does not work with more than 10
-    ## MPI ranks)
-    random_filter_expression = 'numpy.isin(ids + 0.1*cpus,' \
-                                          'ids_filtered_warpx + 0.1*cpus_filtered_warpx)'
-    check_particle_filter(fn, filtered_fn, random_filter_expression)
+dim = "3d"
+species_name = "electron"
 
 parser_filter_fn = "diags/diag_parser_filter00150"
-parser_filter_expression = "(px*py*pz < 0) * (numpy.sqrt(x**2+y**2+z**2)<100)"
-check_particle_filter(last_fn, parser_filter_fn, parser_filter_expression)
+parser_filter_expression = "(px*py*pz < 0) * (np.sqrt(x**2+y**2+z**2)<100)"
+post_processing_utils.check_particle_filter(last_fn, parser_filter_fn, parser_filter_expression,
+                                            dim, species_name)
 
 uniform_filter_fn = "diags/diag_uniform_filter00150"
 uniform_filter_expression = "ids%11 == 0"
-check_particle_filter(last_fn, uniform_filter_fn, uniform_filter_expression)
+post_processing_utils.check_particle_filter(last_fn, uniform_filter_fn, uniform_filter_expression,
+                                            dim, species_name)
 
 random_filter_fn = "diags/diag_random_filter00150"
 random_fraction = 0.88
-check_random_filter(last_fn, random_filter_fn, random_fraction)
+post_processing_utils.check_random_filter(last_fn, random_filter_fn, random_fraction,
+                                          dim, species_name)
 
 test_name = last_fn[:-9] # Could also be os.path.split(os.getcwd())[1]
 checksumAPI.evaluate_checksum(test_name, fn, do_particles=False)
