+++ Regression/WarpX-tests.ini
@@ -619,6 +619,7 @@ compareParticles = 1
 particleTypes = electrons ions
 analysisRoutine = Examples/Tests/Langmuir/analysis_langmuir_multi_rz.py
 analysisOutputImage = Langmuir_multi_rz_analysis.png
+aux1File = Regression/PostProcessingUtils/post_processing_utils.py
 tolerance = 1.e-14
 
 [Langmuir_multi_rz_psatd]
@@ -638,6 +639,7 @@ compareParticles = 1
 particleTypes = electrons ions
 analysisRoutine = Examples/Tests/Langmuir/analysis_langmuir_multi_rz.py
 analysisOutputImage = Langmuir_multi_rz_psatd_analysis.png
+aux1File = Regression/PostProcessingUtils/post_processing_utils.py
 tolerance = 1.e-14
 
 [Langmuir_multi_rz_psatd_current_correction]
@@ -657,6 +659,7 @@ compareParticles = 1
 particleTypes = electrons ions
 analysisRoutine = Examples/Tests/Langmuir/analysis_langmuir_multi_rz.py
 analysisOutputImage = Langmuir_multi_rz_psatd_analysis.png
+aux1File = Regression/PostProcessingUtils/post_processing_utils.py
 tolerance = 1.e-14
 
 [Python_Langmuir_rz_multimode]
@@ -1491,6 +1494,7 @@ compileTest = 0
 doVis = 0
 compareParticles = 0
 analysisRoutine = Examples/Tests/collision/analysis_collision_3d.py
+aux1File = Regression/PostProcessingUtils/post_processing_utils.py
 tolerance = 1.e-14
 
 [collisionXZ]
@@ -1508,6 +1512,7 @@ compileTest = 0
 doVis = 0
 compareParticles = 0
 analysisRoutine = Examples/Tests/collision/analysis_collision_2d.py
+aux1File = Regression/PostProcessingUtils/post_processing_utils.py
 tolerance = 1.e-14
 
 [Maxwell_Hybrid_QED_solver]
